Edit: wetube: done Chapter#8 custom video player.

+++ controllers/videoController.js
@@ -37,9 +37,11 @@ export const postUpload = async (req, res) => {
   const newVideo = await Video.create({
     fileUrl: path,
     title,
-    description
+    description,
+    creator: req.user.id,
   });
-  console.log(newVideo);
+  req.user.videos.push(newVideo.id);
+  req.user.save();
   res.redirect(routes.videoDetail(newVideo.id));
 };
 
@@ -48,7 +50,7 @@ export const videoDetail = async (req, res) => {
     params: { id }
   } = req;
   try {
-    const video = await Video.findById(id);
+    const video = await Video.findById(id).populate('creator');
     res.render("videoDetail", { pageTitle: video.title, video });
   }
   catch (error) {
@@ -62,7 +64,12 @@ export const getEditVideo = async (req, res) => {
   } = req;
   try {
     const video = await Video.findById(id);
-    res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+    if (video.creator !== req.user.id) {
+      throw Error();
+    }
+    else {
+      res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+    }
   }
   catch (error) {
     res.redirect(routes.home);
@@ -88,7 +95,13 @@ export const deleteVideo = async (req, res) => {
     params: { id }
   } = req;
   try {
-    await Video.findOneAndRemove({ _id: id });
+    const video = await Video.findById(id);
+    if (video.creator !== req.user.id) {
+      throw Error();
+    }
+    else {
+      await Video.findOneAndRemove({ _id: id });
+    }
   }
   catch (error) {
     console.log(error);
